# GFBackend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Gator Forum Backend Server",
        "title": "Gator Forum Backend API",
        "termsOfService": "https://github.com/fongziyjun16/SE/tree/backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://167.71.166.120:10010",
    "basePath": "/gf/api",
    "paths": {
...
        "/articletype/all": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Type Manage"
                ],
                "summary": "Get All Article Types",
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.ArticleType"
                            }
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters / Type has existed",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/articletype/create": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need new article type information, cannot repeat type name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Type Manage"
                ],
                "summary": "Create a new article type by admin user",
                "parameters": [
                    {
                        "description": "New Article Type Information",
                        "name": "ArticleTypeInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ArticleTypeInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Create Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters / Type has existed",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/articletype/remove": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need type name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Type Manage"
                ],
                "summary": "Admin user removes article type",
                "parameters": [
                    {
                        "description": "Type Name",
                        "name": "TypeName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Remove Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/articletype/update": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need type name \u0026 new description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Type Manage"
                ],
                "summary": "Admin user update article type description",
                "parameters": [
                    {
                        "description": "Type Name",
                        "name": "TypeName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Description",
                        "name": "Description",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Update Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ArticleType": {
            "type": "object",
            "properties": {
                "create_Day": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "typeName": {
                    "type": "string"
                }
            }
        },
        "entity.ArticleTypeInfo": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string",
                    "example": "Discussion of Movie"
                },
                "TypeName": {
                    "type": "string",
                    "example": "Movie"
                }
            }
        },
...
        "entity.ResponseMsg": {
            "type": "object",
            "properties": {
                "Nickname": {
                    "type": "string",
                    "example": "James Bond"
                },
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "process successfully"
                },
                "new_community_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "cookies"
        }
    }
}